+++ CHANGELOG.md
@@ -21,9 +21,12 @@ We use the following categories for changes:
 - add low-resources preset [#sc-5895]
 - add low-resources flag to `deploy` [#sc-5895]
 - support minikube and kind clusters [#sc-5895]
+- add shepherd resources tune [#sc-6617]
 
 ### Changed
+
 - added a SaaS link on helm validation error [#sc-6297]
+
 ### Fixed
 
 ### Removed